requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-decouple==3.8
python-dotenv==1.0.0
requests==2.29.0
urllib3==1.26.15
Werkzeug==2.3.1
zipp==3.15.0
